Frame generator program
Changing the way the led cube runs.
Modifying it to a more "game loop" style where we are displaying
frames and animations. At each minor tick we change to the next frame
in the animation and on a major tick we change the animation.
make_frames creates a source file that contains all the frame
definitions and pointers to the start of the animation sequences.
The produced file can be included in the ledcube program.